+++ CHANGELOG.md
@@ -6,9 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2024-06-27
 ### Added
 - Random damage generator.
 - Generator of Robe of Useful Items's items.
+## Changed
+- Fully refurbished DDB methods, creating:
+  - Attacks generator;
+  - Traits generator; and
+  - Legendary Actions generator.
 
 ## [1.1.2] - 2022-05-12
 ### Changed
@@ -33,7 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Recharge generation; and
   - General generation.
 
-[Unreleased]: https://github.com/Nereare/DDB-Rollable/compare/v1.1.2..HEAD
+[Unreleased]: https://github.com/Nereare/DDB-Rollable/compare/v2.0.0..HEAD
+[2.0.0]: https://github.com/Nereare/DDB-Rollable/compare/v1.1.2...v2.0.0
 [1.1.2]: https://github.com/Nereare/DDB-Rollable/compare/v1.1.0...v1.1.2
 [1.1.0]: https://github.com/Nereare/DDB-Rollable/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/Nereare/DDB-Rollable/releases/tag/v1.0.0